yt_script_stack/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
requests==2.32.3
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
